+++ apps/api/src/__tests__/routes/fields/archive.test.ts
@@ -13,13 +13,8 @@ import type { ApiErrorResponse } from "@pluralscape/types";
 
 // ── Mocks ────────────────────────────────────────────────────────
 
-vi.mock("../../../services/field-definition.service.js", () => ({
-  listFieldDefinitions: vi.fn(),
-  createFieldDefinition: vi.fn(),
-  getFieldDefinition: vi.fn(),
-  updateFieldDefinition: vi.fn(),
+vi.mock("../../../services/field-definition/archive.js", () => ({
   archiveFieldDefinition: vi.fn(),
-  restoreFieldDefinition: vi.fn(),
 }));
 
 vi.mock("../../../lib/audit-writer.js", () => mockAuditWriterFactory());
@@ -33,7 +28,7 @@ vi.mock("../../../lib/system-ownership.js", () => mockSystemOwnershipFactory());
 vi.mock("../../../middleware/auth.js", () => mockAuthFactory());
 // ── Imports after mocks ──────────────────────────────────────────
 
-const { archiveFieldDefinition } = await import("../../../services/field-definition.service.js");
+const { archiveFieldDefinition } = await import("../../../services/field-definition/archive.js");
 const { systemRoutes } = await import("../../../routes/systems/index.js");
 
 // ── Helpers ──────────────────────────────────────────────────────
