+++ apps/api/src/routes/fields/create.ts
@@ -9,7 +9,7 @@ import { parseJsonBody } from "../../lib/parse-json-body.js";
 import { envelope } from "../../lib/response.js";
 import { createIdempotencyMiddleware } from "../../middleware/idempotency.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { createFieldDefinition } from "../../services/field-definition.service.js";
+import { createFieldDefinition } from "../../services/field-definition/create.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
