+++ apps/api/src/routes/fields/get.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { getFieldDefinition } from "../../services/field-definition.service.js";
+import { getFieldDefinition } from "../../services/field-definition/get.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
