+++ apps/api/src/services/field-definition/archive.ts
@@ -1,0 +1,59 @@
+import { fieldDefinitions } from "@pluralscape/db/pg";
+import { now } from "@pluralscape/types";
+import { and, eq } from "drizzle-orm";
+
+import { HTTP_NOT_FOUND } from "../../http.constants.js";
+import { ApiHttpError } from "../../lib/api-error.js";
+import { withTenantTransaction } from "../../lib/rls-context.js";
+import { assertSystemOwnership } from "../../lib/system-ownership.js";
+import { tenantCtx } from "../../lib/tenant-context.js";
+
+import { invalidateFieldDefCache } from "./internal.js";
+
+import type { AuditWriter } from "../../lib/audit-writer.js";
+import type { AuthContext } from "../../lib/auth-context.js";
+import type { FieldDefinitionId, SystemId } from "@pluralscape/types";
+import type { PostgresJsDatabase } from "drizzle-orm/postgres-js";
+
+export async function archiveFieldDefinition(
+  db: PostgresJsDatabase,
+  systemId: SystemId,
+  fieldId: FieldDefinitionId,
+  auth: AuthContext,
+  audit: AuditWriter,
+): Promise<void> {
+  assertSystemOwnership(systemId, auth);
+
+  await withTenantTransaction(db, tenantCtx(systemId, auth), async (tx) => {
+    const [existing] = await tx
+      .select({ id: fieldDefinitions.id })
+      .from(fieldDefinitions)
+      .where(
+        and(
+          eq(fieldDefinitions.id, fieldId),
+          eq(fieldDefinitions.systemId, systemId),
+          eq(fieldDefinitions.archived, false),
+        ),
+      )
+      .limit(1);
+
+    if (!existing) {
+      throw new ApiHttpError(HTTP_NOT_FOUND, "NOT_FOUND", "Field definition not found");
+    }
+
+    const timestamp = now();
+
+    await audit(tx, {
+      eventType: "field-definition.archived",
+      actor: { kind: "account", id: auth.accountId },
+      detail: "Field definition archived",
+      systemId,
+    });
+
+    await tx
+      .update(fieldDefinitions)
+      .set({ archived: true, archivedAt: timestamp, updatedAt: timestamp })
+      .where(and(eq(fieldDefinitions.id, fieldId), eq(fieldDefinitions.systemId, systemId)));
+  });
+  invalidateFieldDefCache();
+}
