+++ apps/api/src/services/field-definition/get.ts
@@ -1,0 +1,44 @@
+import { fieldDefinitions } from "@pluralscape/db/pg";
+import { and, eq } from "drizzle-orm";
+
+import { HTTP_NOT_FOUND } from "../../http.constants.js";
+import { ApiHttpError } from "../../lib/api-error.js";
+import { withTenantRead } from "../../lib/rls-context.js";
+import { assertSystemOwnership } from "../../lib/system-ownership.js";
+import { tenantCtx } from "../../lib/tenant-context.js";
+
+import { toFieldDefinitionResult } from "./internal.js";
+
+import type { FieldDefinitionResult } from "./internal.js";
+import type { AuthContext } from "../../lib/auth-context.js";
+import type { FieldDefinitionId, SystemId } from "@pluralscape/types";
+import type { PostgresJsDatabase } from "drizzle-orm/postgres-js";
+
+export async function getFieldDefinition(
+  db: PostgresJsDatabase,
+  systemId: SystemId,
+  fieldId: FieldDefinitionId,
+  auth: AuthContext,
+): Promise<FieldDefinitionResult> {
+  assertSystemOwnership(systemId, auth);
+
+  return withTenantRead(db, tenantCtx(systemId, auth), async (tx) => {
+    const [row] = await tx
+      .select()
+      .from(fieldDefinitions)
+      .where(
+        and(
+          eq(fieldDefinitions.id, fieldId),
+          eq(fieldDefinitions.systemId, systemId),
+          eq(fieldDefinitions.archived, false),
+        ),
+      )
+      .limit(1);
+
+    if (!row) {
+      throw new ApiHttpError(HTTP_NOT_FOUND, "NOT_FOUND", "Field definition not found");
+    }
+
+    return toFieldDefinitionResult(row);
+  });
+}
